# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/asset-ms/v1/web/assets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "List assets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "USD",
                        "description": "Currency",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status active|inactive",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/data.AssetVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/asset-ms/v1/web/assets/{asset_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get asset detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "USD",
                        "description": "Currency",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.AssetVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/asset-ms/v1/web/holdings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holding"
                ],
                "summary": "List holdings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.HoldingListVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/asset-ms/v1/web/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pending|pay_succeed|pay_fail",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Unix timestamp",
                        "name": "created_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Unix timestamp",
                        "name": "created_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Last seen order_id",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.OrderListVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.OrderVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/asset-ms/v1/web/orders/{order_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get order detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.OrderDetailVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/asset-ms/v1/web/quotes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quote"
                ],
                "summary": "Generate quote",
                "parameters": [
                    {
                        "description": "Quote request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.GenQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.QuoteVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.AssetVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "current_price": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "data.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "err_msg": {
                    "type": "string"
                }
            }
        },
        "data.CreateOrderRequest": {
            "type": "object",
            "required": [
                "idempotency_key",
                "payment_method_id",
                "quantity",
                "quote_id"
            ],
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                }
            }
        },
        "data.GenQuoteRequest": {
            "type": "object",
            "required": [
                "asset_id"
            ],
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "data.HoldingAssetVO": {
            "type": "object",
            "properties": {
                "icon_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "data.HoldingListVO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.HoldingVO"
                    }
                }
            }
        },
        "data.HoldingVO": {
            "type": "object",
            "properties": {
                "asset": {
                    "$ref": "#/definitions/data.HoldingAssetVO"
                },
                "asset_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "valuation": {
                    "$ref": "#/definitions/data.HoldingValuationVO"
                }
            }
        },
        "data.HoldingValuationVO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
        "data.OrderAssetVO": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "data.OrderDetailVO": {
            "type": "object",
            "properties": {
                "asset": {
                    "$ref": "#/definitions/data.OrderAssetVO"
                },
                "created_at": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string"
                },
                "pay_amount": {
                    "type": "string"
                },
                "pay_currency": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "data.OrderListVO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.OrderVO"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "data.OrderVO": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string"
                },
                "pay_amount": {
                    "type": "string"
                },
                "pay_currency": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "data.QuoteVO": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "integer"
                },
                "quote_id": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        }
    }
}